add useUserStreams
